tendencias20252/requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2.6
django-filter==25.1
django-import-export==4.3.10
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-spectacular==0.28.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
numpy==2.3.3
oauthlib==3.3.1
pandas==2.3.3
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0